% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBtuning.R
\name{checkargs.xy}
\alias{checkargs.xy}
\title{Estimate the variance of a linear regression}
\usage{
checkargs.xy(x, y)
}
\arguments{
\item{Input_X}{Predictors}

\item{Input_Y}{Outcome}
}
\description{
This estimates the variance of a linear regression by SI
}
\seealso{
\code{\link[glmnet]{cv.glmnet}}
}
