% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.ipreg}
\alias{coef.ipreg}
\title{Extract model coefficients from fitted \code{ipreg} object}
\usage{
\method{coef}{ipreg}(object, ...)
}
\arguments{
\item{object}{Fitted 'ipreg' model object.}

\item{...}{Not used}
}
\value{
Coefficients extracted from the fitted model.
}
\description{
\code{coef.ipreg} extracts model coefficients from objects returned by \code{ipreg} object.
}
\details{
\code{coef} and \code{predict} methods are provided as a convenience to extract coefficients and make prediction. \code{coef.ipreg} simply extracts the estimated coefficients returned by \code{ipreg}.
}
\examples{
## see examples in predict.ipreg()
}
\seealso{
\code{ipreg}, \code{predict.ipreg}
}
