% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.xtune.R
\name{coef_xtune}
\alias{coef_xtune}
\title{Extract model coefficients from fitted \code{xtune} object}
\usage{
coef_xtune(object, ...)
}
\arguments{
\item{object}{Fitted 'xtune' model object.}

\item{...}{Not used}
}
\value{
Coefficients extracted from the fitted model.
}
\description{
\code{coef_xtune} extracts model coefficients from objects returned by \code{xtune} object.
}
\details{
\code{coef} and \code{predict} methods are provided as a convenience to extract coefficients and make prediction. \code{coef.xtune} simply extracts the estimated coefficients returned by \code{xtune}.
}
\examples{
# See examples in \code{predict_xtune}.
}
\seealso{
\code{xtune}, \code{predict_xtune}
}
