% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cus_ridge.R
\name{cus_ridge}
\alias{cus_ridge}
\title{Customized LASSO using iterative reweighted l2}
\usage{
cus_ridge(X, Y, Z, sigma.square = estimateVariance(X, Y),
  alpha.init = rep(0, ncol(Z)), maxstep = 100, margin = 0.001,
  maxstep_inner = 50, tol.inner = 0.1, compute.likelihood = FALSE,
  verbosity = 1)
}
\arguments{
\item{X}{predictor matrix of dimension \eqn{n*q} used for previous regression}

\item{Y}{continuous outcome vector of dimension \eqn{p}.}

\item{Z}{external information data matrix of dimension \eqn{p*q}}

\item{sigma.square}{variance estimation, default is the estimated variance using R package "selectiveinference"}

\item{alpha.init}{initial value for alpha, default is rep(1,p)}

\item{maxstep}{maximal step of iterations}

\item{margin}{stoping creteria for converge}

\item{maxstep_inner}{maximum inner loop step}

\item{compute.likelihood}{compute likelihood or not}

\item{verbosity}{track update process or not}

\item{y}{continuous outcome vector of dimension \eqn{n} used for previous regression}

\item{beta_hat}{estimated beta coefficients from regression (no intercept term)}

\item{x_predict}{data for prediction (no column of 1 needed)}

\item{delta.init}{initial value for delta, the auxilary function for updating alpha}

\item{tol.delta}{stopping creteria for inner loop}
}
\value{
Returns the estimated beta coefficients for classification and pseudo probablity
}
\description{
\code{cus_ridge} estimated beta coefficients and pseudo probablity
}
