% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomizedLASSO_reweighted_l2.R
\name{CusLasso.l2}
\alias{CusLasso.l2}
\title{Customized LASSO using iterative reweighted l2}
\usage{
CusLasso.l2(X, Y, Z, sigma.square = estimateVariance(X, Y),
  alpha.init = rep(1, ncol(Z)), delta.init = rep(1, ncol(X)),
  maxstep = 100, margin = 0.001, maxstep_inner = 100, tol.delta = 0.001,
  compute.likelihood = FALSE)
}
\arguments{
\item{X}{predictor matrix of dimension \eqn{n*q} used for previous regression}

\item{Y}{continuous outcome vector of dimension \eqn{p}.}

\item{Z}{external information data matrix of dimension \eqn{p*q}}

\item{sigma.square}{variance estimation, default is the estimated variance using R package "selectiveinference"}

\item{alpha.init}{initial value for alpha, default is rep(1,p)}

\item{delta.init}{initial value for delta, the auxilary function for updating alpha}

\item{maxstep}{maximal step of iterations}

\item{margin}{stoping creteria for converge}

\item{maxstep_inner}{maximum inner loop step}

\item{tol.delta}{stopping creteria for inner loop}

\item{compute.likelihood}{compute likelihood or not}

\item{y}{continuous outcome vector of dimension \eqn{n} used for previous regression}

\item{beta_hat}{estimated beta coefficients from regression (no intercept term)}

\item{x_predict}{data for prediction (no column of 1 needed)}

\item{X}{predictor matrix of dimension \eqn{n*q}.}
}
\value{
Returns the estimated beta coefficients for classification and pseudo probablity
}
\description{
\code{CusLasso.l2} estimated beta coefficients and pseudo probablity
}
\examples{
CusLassol2.result = CusLasso.l2(x_train,y_train,Z,sigma.square = sig.est,compute.likelihood = T)
}
