% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_EBtuning.R
\name{eb_tuning}
\alias{eb_tuning}
\title{Empirical Bayes estimation of Lasso tuning parameter.}
\usage{
eb_tuning(input_X, input_Y, initial_val = 0.1, maxstep = 100,
  margin = 0.01)
}
\arguments{
\item{input_X}{predictor matrix of dimension \eqn{n*q}.}

\item{input_Y}{continuous outcome vector of dimension \eqn{p}.}

\item{initial_val}{initial value for \eqn{\tau}. Default value is 0.1.}

\item{margin}{convergence criteria. Default value is 0.01.}

\item{max_step}{max step for iterations. Default value is 100.}
}
\value{
Returns the estimated tuning parameter
}
\description{
\code{eb_tuning} returns the tuning parameter estimated by empirical bayes method.
}
