% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{example}
\alias{example}
\title{An simulated example dataset}
\format{The \code{example} object is a list containing three elements:
\itemize{
\item X: A simulated 100 by 200 matrix
\item Y: Continuous response vector of length 100
\item Z: A 200 by 3 matrix. Z_jk indicates whether predictor X_j has external variable Z_k or not.
}}
\usage{
data(example)
}
\description{
The simulated \code{example} data contains 100 observations, 200 predictors, and an continuous outcome. Z contains 3 columns, each column is
indicator variable (can be viewed as the grouping of predictors). The outcome Y is continuous.
}
\examples{
data(example)
X <- example$X
Y <- example$Y
Z <- example$Z
\donttest{ipreg(X,Y,Z)}
}
\keyword{datasets}
