% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misclassification.R
\name{misclassification}
\alias{misclassification}
\title{Calculate misclassification error}
\usage{
misclassification(pred, true)
}
\arguments{
\item{pred}{Predicted class}

\item{true}{Actual class}
}
\value{
Misclassification error for binary or multiclass outcome.
}
\description{
\code{misclassification} calculate misclassification error between predicted class and true class
}
\examples{
Y1 <- rbinom(10,1,0.5)
Y2 <- rnorm(10,1,0.5)
misclassification(Y1,Y2)
}
\seealso{
To calculate the Area Under the Curve (AUC) for binary or multiclass outcomes, please refer to the \code{pROC}.
}
