% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\title{Calculate mean square error}
\usage{
mse(pred, true)
}
\arguments{
\item{pred}{Prediction values vector}

\item{true}{Actual values vector}
}
\value{
mean square error
}
\description{
\code{mse} calculate mean square error (MSE) between prediction values and true values
}
\examples{
Y1 <- rnorm(10,0,1)
Y2 <- rnorm(10,0,1)
mse(Y1,Y2)
}
