% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_to_classify.R
\name{reg.to.classify}
\alias{reg.to.classify}
\title{High dimensional LDA to build classifyer using regression results}
\usage{
reg.to.classify(x, y, beta_hat, x_predict)
}
\arguments{
\item{x}{predictor matrix of dimension \eqn{n*q} used for previous regression}

\item{y}{continuous outcome vector of dimension \eqn{n} used for previous regression}

\item{beta_hat}{estimated beta coefficients from regression (no intercept term)}

\item{x_predict}{data for prediction (no column of 1 needed)}
}
\value{
Returns the estimated beta coefficients for classification and pseudo probablity
}
\description{
\code{reg.to.classify} estimated beta coefficients and pseudo probablity
}
\examples{
reg.to.classify(train_X,train_Y,cusLasso.result$coef,testX)
}
