% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb_tuning.R
\name{eb_tuning}
\alias{eb_tuning}
\title{Empirical Bayes estimation of Lasso tuning parameter, add ridge future}
\usage{
eb_tuning(X, Y, var.ini = estimateVariance(X, Y), initial_val = 0.1,
  maxstep = 100, margin = 0.01, verbosity = 0, var.compare = TRUE,
  var.fix = FALSE)
}
\arguments{
\item{X}{predictor matrix of dimension \eqn{n*q}.}

\item{Y}{continuous outcome vector of dimension \eqn{p}.}

\item{var.ini}{initial estimates of variance from external estimation}

\item{initial_val}{initial value for \eqn{\tau}. Default value is 0.1.}

\item{maxstep}{max step for iterations. Default value is 100.}

\item{margin}{convergence criteria. Default value is 0.01.}

\item{verbosity}{whether to track the update process}

\item{var.compare}{whether or not to compare the variance estimates from different methods}

\item{var.fix}{if TRUE, fix variance estimates}
}
\value{
Returns the estimated tuning parameter and estimated coefficients
}
\description{
\code{eb_tuning} returns the tuning parameter estimated by empirical bayes method.
}
