% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_variance.R
\name{estimateVariance}
\alias{estimateVariance}
\title{Estimate Variance given predictor X and response Y.}
\usage{
estimateVariance(X, Y, num = 10)
}
\arguments{
\item{X}{predictor matrix of dimension \eqn{n*q} used for previous regression}

\item{Y}{continuous outcome vector of dimension \eqn{n} used for previous regression}

\item{num}{number of repeated estimation}
}
\description{
\code{estimateVariance} estimate noise variance
}
