% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipreg.R
\name{ipreg}
\alias{ipreg}
\title{L1 and L2 penalized regression with individual amount of shrinkage for each regression coefficient based on external information.}
\usage{
ipreg(X, Y, Z = NULL, sigma.square = NULL, method = c("lasso", "ridge"),
  control = list())
}
\arguments{
\item{X}{predictor matrix of dimension \eqn{n*q} used for previous regression}

\item{Y}{continuous outcome vector of dimension \eqn{n} used for previous regression}

\item{Z}{external information data matrix of dimension \eqn{p*q}}

\item{sigma.square}{variance estimation, default is the estimated variance using R package "selectiveinference"}

\item{method}{method = "Lasso" for Lasso regression, method = "Ridge" for Ridge regression}

\item{control}{specifies ipreg control object. See \code{\link{ipreg.control}} for more details.}
}
\description{
\code{ipreg} estimated beta coefficients and pseudo probablity
}
