% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ipreg}
\alias{predict.ipreg}
\title{Predict model for integrated penalized regression model fit}
\usage{
\method{predict}{ipreg}(object, newX, ...)
}
\arguments{
\item{object}{Fitted "ipreg" model object.}

\item{newX}{Matrix of values at which predictions are to be made.}

\item{...}{Not used}
}
\description{
\code{predict.ipreg} produces predicted values fitting an ipreg model to a new dataset
}
