% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.xtune}
\alias{predict.xtune}
\title{Model predictions based on fitted \code{xtune} object}
\usage{
\method{predict}{xtune}(object, newX, type = c("response", "class"), ...)
}
\arguments{
\item{object}{Fitted 'xtune' model object.}

\item{newX}{Matrix of values at which predictions are to be made.}

\item{...}{Not used}
}
\value{
A vector of predictions
}
\description{
\code{predict.xtune} produces predicted values fitting an xtune model to a new dataset
}
\details{
\code{coef} and \code{predict} methods are provided as a convenience to extract coefficients and make prediction. \code{predict.xtune} simply calculate the predicted value using the estimated coefficients returned by \code{xtune}.
}
\examples{
## simulate data
set.seed(9)
data(example)
X <- example$X
Y <- example$Y
Z <- example$Z

## If no Z provided, perform Empirical Bayes tuning
fit.eb <- xtune(X,Y)
## Coef and predict methods
coef(fit.eb)
predict(fit.eb,X)

## Differential shrinkage based on external information Z:
fit.diff <- xtune(X,Y,Z)
## Coef and predict methods
coef(fit.diff)
predict(fit.diff,X)
}
\seealso{
\code{xtune}, \code{coef.xtune}
}
